Rename Councilor file to match interface
