+++ packages/ui/src/council/types/index.ts
@@ -1,1 +1,1 @@
-export * from './Concilor'
+export * from './Councilor'
